# exec: optimize builder in vectorized merge joiner